ENH: Added NLOptPrecond
For nonlinear optimisers, the preconditioner (if any) is not static but
needs to be re-evaluated after each update of the state variable. The
new NLOptPrecond class offers a prepare() function similar to the one
used in the nonlinear functionals already (i.e. in NLCG, NLSD and the
ALGLIB wrappers).

This commit also adds implementations of this for the
HessianPreconditions using AnalyticFunctionOperators.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>