ENH: Added reset() function to Linesearches
Usually a Linesearch object is reused by a nonlinear optimiser and
sometimes needs to reset some member variables. Most notably in the
StrongWolfeLinesearch, where you usually want to keep the length of the
last good step not only throughout all iterations of the Linesearch, but
also through all iterations of the nonlinear optimiser using that
Linesearch. When reusing the outer solver, resetting the Linesearch is a
good idea.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>