revised trafo and space configuration tags
The Trafo::Config and Space::Config classes have been replaced by
corresponding enums with operator overloads to simplify the specification
of the required evaluation quantities. This should help to avoid erroneous
configurations that can be produced by simple typos.

Moreover, this commit also simplifes the trafo and space evaluation data
classes, which are used internally by the assembly, by removing some of
the wicked template magic tricks.

Although this commit modifies a lot of files, it is mostly a
"under-the-hood" commit.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>