merged and simplified SymbolicGraph/MatrixAssembler class
The SymbolicGraphAssembler and SymbolicMatrixAssembler class templates
have been merged into a single (non-template) SymbolicAssembler class. The
previous splitting and templatisation in the stencil was somewhat
"artistic" in the sense that it did not yield any practicable gain in
comparison to the new implementation, which simply offers the various
supported sparsity patterns by various static member functions.

The changes to the user code are mostly cosmetrical, e.g. for the assembly
of a standard matrix pattern one now calls

Assembly::SymbolicAssembler::assemble_matrix_std1/2(...)

instead of

Assembly::SymbolicMatrixAssembler<>::assemble_matrix1/2(...)

The main reason for this change is code simplification.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>