removed obsolete CONTEXT macro
This episode of this year's "Fruehjahrsputz" removes the obsolete CONTEXT
macro, which was used internally by our exception system to implement
something similar to a call-stack trace. However, kernel developers have
stopped using the CONTEXT macro for a long while now, so it has become
pointless over the time.

In consequence, all occurances of the CONTEXT macro as well as the
context-stack-tracing functionality of the Exception class, which was used
by the CONTEXT macro, have been removed.

Most probably the ASSERT macro and its internal implementation as an
Exception will face a major revisement soon, too, with the aim of
providing some sort of call-stack in case of a failed assertion.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>