ENH: Dimension-dependent helper class for parsing charts
Until now, there was no need to skip certain chart types based on the
world dimension, but with the recent changes, GCC 6.1 rightfully
complained that there might be a pathologic case of uninitialised
memory. The clean solution is that chart types are only checked for if
they are appropriate for the world dimension of the mesh they are
supposed to work with.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>