FIX: Various nonlinear solver fixes
The trim threshold parameter in the linesearch needs to be computed once
per outer solver application, not in every invocation of the linesearch.

Logic fixes in the NLCG restart.

The tests now have a hard-coded acceptable maximum number of iterations.
In the past, new errors I introduced into the solvers went unnoticed
because they still converged, but needed more iterations.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>