renamed vectors in PCG solver to conform with Saad'03
Moreover, added some useful documentation of the PCG internals. The
vectors have been renamed to conform to the famous Saad book, which is a
widely known and accepted reference when it comes to iterative solvers.

Note that this is a purely cosmetical commit which does not change the
semantics of the PCG solver.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>