added PCGNE and PCGNEILU solvers
These two solver classes represent a (preconditioned) Conjugate-Gradient
solver applied onto the normal equation system:

   (A^T * A) * x = A^T * b

The PCGNE solver implements the CG algorithm by using to (possibly
different) SPD preconditioners M_L and M_R. Unfortunately, such SPD
preconditioners are hard to come by if the system matrix A itself is not
SPD, so this solver class is mostly of academic interest, as it requires
special preconditioners, which are not (yet) available in FEAT.

The PCGNEILU solver implements a special PCGNE (see above) variant which
uses the ILU(p) factorisation of A to obtain a pair of SPD preconditioners

- M_L := L * L^T
- M_R := U^T * U

where L and U are the triangular factor matrices of the ILU(p)
factorisation.

Note that both solver classes are experimental and should be used with
care.

Moreover, note that both of the new solver classes solve a normal equation
system, which has a condition number of O(h^{-4}) for second order
elliptic PDEs -- in contrast to O(h^{-2}) for the original system.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>