added Windows OS utility functions
This commit adds a set of OS utility functions for Windows. This was
necessary as the Windows headers do not mix well with FEAT headers and
compiler settings due to various name collisions and other nasty
conflicts. All previous calls to the Win32 API have been encapsulated in
the new FEAT::Windows namespace.

Moreover, this commit also changes the error handling in the FEAT::Runtime
functions, which now write directly to std::cerr and call abort() instead
of throwing an InternalError.

Next, the Runtime::initialise now takes care of disabling error prompts
for the automated VC test system. This is necessary as otherwise the test
system would block awaiting user interaction.

Finally, the Runtime::abort() function now has a parameter which specifies
whether the abort function should dump the call-stack to stderr prior to
terminating the process. Note that this functionality is currently only
available on Windows; support for other platforms will be added soon.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>