revised ASSERT implementation
The previous implemenation of the ASSERT and ASSERT_ macros used a custom
exception class "Assertion" that was thrown if an assertion failed. This
was somewhat fishy, as an exception is a "possibly-recoverable" error
that can be handled by a surrounding try-catch block, whereas an assertion
is an unrecoverable and therefore fatal error.

The core of the new implementation is a function named "assertion", which
handles the output of an appropriate error message in case of a failed
assertion. This function is not called directly, but instead by using one
of four assertion macros ASSERT, ASSERTM, XASSERT and XASSERTM.

The differences between the four macros are:

1) ASSERT and ASSERTM only call the assertion function when the
   pre-processor macro DEBUG is defined, and are defined as empty macros
   otherwise. This corresponds to the definition of the Standard-C
   "assert" macro.

2) XASSERT and XASSERTM are always defined to call the assertion function,
   i.e. independent of whether DEBUG is defined or not.

3) ASSERTM and XASSERTM have a second argument of type "const char*const"
   in addition to the asserted boolean expression, which is a user-defined
   error message that is also written to stderr in case that the assertion
   fails. Note that combining an error message using the FEAT::String
   class and/or the stringify functions is discouraged, as this string
   arithmetic may lead to errors by itself.

As a rule of thumb, you should obey the following guidelines to choose the
appropriate macro in your current situation:

1) If the assertion is in a performance-critical function, you should use
   the ASSERT or ASSERTM macro to avoid runtime overhead in non-debug
   builds, which would be introduced by the XASSERT or XASSERTM macros.

2) If you have something informative to say about the assertion, you
   should use the ASSERTM or XASSERTM macro to provide a custom error
   message. Note that the stringified expression is always printed to
   stderr, so you don't need to specify the expression as a error message
   explicitly anymore.

The assertion function, which is called by all four of the above macros,
always writes the following information to stderr if the assertion fails:

1) A line "ERROR: ASSERTION FAILED"
   or "ERROR: ASSERTION FAILED: <msg>",
   if the ASSERTM or XASSERTM macro was used

2) The stringified failed expression

3) The name of the function that contains the failed assertion.

4) Filename and line number of the failed assertion.

Moreover, the assertion function calls Runtime::abort in case of a failed
assertion to terminate the calling process (and MPI world, if applicable)
immediately. The abort function itself then dumps the call-stack to
stderr (if possible) to provide even more information about the failed
assertion.

All instances of the "old" ASSERT and ASSERT_ macro calls have been
replaced by one of the four new macros, whichever seemed to be the most
appropriate for the situation by obeying the guidelines above.

Note that many instances of "throw InternalError(...)" may now be replaced
by corresponding calls to XASSERT or XASSERTM.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>