added call-stack back-trace for linux systems
The current implementation is rather simple as 1) it does not demangle
the function names and 2) it does not map addresses to filenames and
line numbers. If anyone is bored, (s)he may take a look at the
following sites for more information:

http://blog.bigpixel.ro/2010/09/stack-unwinding-stack-trace-with-gcc
http://stackoverflow.com/questions/6934659/how-to-make-backtrace-backtrace-symbols-print-the-function-names

It seems that using the 'addr2line' tool from binutils is one way to get
the filename/line-number mapping job done. Moreover, according to Dirk
(and various online sources) the tool 'c++filt' is recommended for the
demangling of function names.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>