ENH: Added Hager-Zhang direction update to NLCG
This update behave different from all others implemented so far and
seems to require more precision in the line search. As I do not want to
implement the line search from the given reference, it remains inferior
to most other direction updates and is meant as a starting point if
someone wants to implement this other line search.

Also changed some variable names in NLCG and NLSD to resemble the ones
in PCG, which are named according to Saad 03.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>