added preconditioned Steepest-Descent and Minimal-Resiual iterations
This commit adds the two well-known preconditioned descent methods.

The preconditioned Steepest-Descent (PSD) solver is the little sister of
the famous PCG solver. PSD performs the same minimisation as PCG, but it
does not orthogonalise the descent direction vectors. Moreover, the PSD
method requires a SPD matrix as well as a SPD preconditioner.

The preconditioned Minimal-Residual (PMR) solver is the little sister of
the not-so-famous PCR solver -- at least to my understanding. PMR performs
the same minimisation as PCR, but does not orthogonalise the descent
direction vectors.

According to Saad's book, the PMR requires the system matrix to be positive
definite, but not necessarily symmetric. This is somewhat confusing, as
the PCR method requires the system matrix to be symmetric but not
necessarily positive definite... (?)

Note that the preconditioner of the PMR solver is assumed to be SPD,
although the system matrix does not need to be symmetric.

Important:
The PMR solver is a "Minimal Residual Iteration" (MR), which is NOT to be
confused with the "Minimal Residual Method" solver (MINRES). Despite the
virtually identical name, these two solvers are actually non-equivalent
algorithms.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>