implemented multigrid preconditioner
The provisional "BasicVCycle" preconditioner got superseeded by the new
MultiGridHierarchy and MultiGrid class pair, which provide a more
sophisticated implementation of a multigrid solver.

One major difference between the new and the old implementation is that
one now creates an object of type MultiGridHierarchy, which is responsible
for managing the multigrid hierarchy which holds (references/pointers to)
the following data for each level:

* The system matrix and filter
* Grid transfer operators
* Coarse-Grid-Solver
* Smoothers
* Temporary vectors

Then, one can easily create a MultiGrid preconditioner object by using the
Solver::new_multigrid function, which takes a reference to the multigrid
hierarchy as its parameter.

Thanks to this "splitting" between the actual hierarchy and the multigrid
preconditioner, it is possible to create multiple multigrid instances
representing (possibly different ranges of) the same multigrid hierarchy,
thus using the same smoothers and other sub-objects. This will be an
important feature required for building ScaRC-like solvers.

Moreover, this new implementation now also supports the F- and W-cycles in
addition to the V-cycle.

Tutorial 05 has been adapted to use these new MultiGrid[Hierarchy]
classes, but all other applications have not yet been ported.

Note that there are still some minor open questions and ToDos in this new
implementation, which have to be addressed in future. Most of these
"issues" are related to the collection of statistics.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>