ENH: Computation of signed distance to charts
It is now possible to compute the signed distance of a point to a chart,
where the distance is positive if the point lies in the direction of the
interior normal ("inside", or "left" of an oriented edge) and negative
if it lies in the direction of the outward facing normal.

At the moment, this can be accessed through a virtual function of
ChartBase, which is not very efficient due to possible vtable lookup
overhead.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>