implemented varying-degree Spline chart
This new chart class implements a 2D spline of varying degree. This chart
is the successor of the Polyline (degree=1) and Bezier (degree=3) charts,
as both of these are just "boring" special cases. The maximum allowed
degree for the Spline chart has been limited to 5, because a) higher
degree is not really useful and b) point projection would become quite
costly otherwise.

The Polyline and Bezier charts still remain for now, mostly for debugging
purposes. Also, no mesh files have been ported from Polyline/Bezier charts
up to now -- this will be done step by step later. Once all mesh files
have been ported, we may remove the Polyline and Bezier charts to avoid
redundancy.

Note that the two crucial chart functions, namely 'map_on_segment' and
'project_on_segment', explicitly check whether a segment represents a
simple line and, if so, perform a special case handling, so that no
computational overhead is performed in this case. In other words: using a
Spline chart instead of a Polyline chart should not yield any negative
effects w.r.t performance.

The main reason for why a new chart class had to be implemented is that
some domains (e.g. the nozzle or venturi pipe) have both curved and linear
segments. Unfortunately, representing a line segment as a *parametrised*
Bezier curve is not as simple as it may seem, so that a Spline chart of
varying degree was necessary.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>