StrongWolfeLinesearch rewrite
Before, it did not keep the upper limit of the interval of uncertainty,
which is needed for some cases.

The main routine body is no longer split into a seperate routine that
determines the interval and the bracketing iteration, but rather a
single routine using flags.

This all means it is now closer in logic to MINPACKs MCSRCH and LINMIN_MCSTEP
routines than before, but it differs in many implementational details.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>