added TimeStamp::elapsed[_micros|_string]_now() functions
This commit adds three new functions to the TimeStamp class:
* elapsed_now()
* elapsed_micros_now()
* elapsed_string_now()

Calling one of these functions on a TimeStamp object 'ts' is equivalent to
calling the corresponding "not-now" function for a temporary TimeStamp
object with 'ts' as the 'before' argument, e.g.:

    ts.elapsed_now() <==> TimeStamp().elapsed(ts)

These functions should reduce the total number of time-stamp objects
required by a few.

Note that these new functions do not stamp their object, i.e. the internal
time stamp value of the object is left unmodified.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>