workaround for weird MSVC error
The MSVC compiler complained about the declaration of a static-constexpr
variable by emitting an (officially undocumented) error C2131 when
compiling in debug mode on x86, stating that the expression contained an
undefined result (altough in fact, 1/4 = 0.25 is welldefined).

As it seems, this is some flaw in the constexpr engine of the compiler,
which has been circumvented by replacing 'constexpr' by 'const', as it
does not hurt in this particular case.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>