implemented Burgers assembler class
This class is used for the assembly of the non-linear Burgers operator,
which is part of the combined operator of the Navier-Stokes equations.

Currently, this assembly class is quite limited in its capabilities.
However, it is capable of assembling the most basic Burgers variants,
which are required for the most basic Navier-Stokes solvers.

Note that this class is currently only implemented for SparseMatrixBCSR
and DenseVectorBlocked.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>