added Navier-Stokes CP2D CN-Q2/Q1 toy-code application
This commit adds the first non-stationary non-linear meta-blocked
application, which is a "simple" 2D Q2/Q1 Navier-Stokes solver following
the "CP" (Coupled solver, Projection preconditioner) approach on a
equidistant Crank-Nicolson time discretisation with fully implicit
pressure treatment.

Note that this application is a "toy-code", i.e. it is more a
proof-of-concept and a playground for the HPC guys rather than a serious
application -- the latter one will be implemented in near future.

This application is restricted to the "flow-through-a-2D-domain" scenario
with parabolic inflow on one side of the domain and "do-nothing" outflow
condition on the other side of the domain.

The actual problem to be solved can be configured by a set of command line
arguments, which are documented within the application source code.

Moreover, the application supports three pre-configured problems, which
can be easily selected by using the "--setup <config>" option.

The pre-configured problems are:

1) Poiseuille-Flow on the Unit-Square
   The most simple problem, whose solution converges to a steady state
   over time.
   Select via: "--setup square"

2) Jet-Flow through a Nozzle
   Basically the same as the previous one, but on a more interesting
   domain. Also converges to a steady state solution.
   Select via: "--setup nozzle"

3) Flow Around a Cylinder (TM)
   If you don't know what that is, you may (re)consider looking for
   another job...  ;) Jokes aside: That's the non-steady variant for
   Re=100, which generates the famous periodic von-Karman vortex street.
   Select via: "--setup bench1"

The configuration of the non-linear CP-solver is very limited to keep the
code simple. In its pre-defined configuration, the solver approach is:

1) Perform 1 fully-coupled non-linear iteration per time-step. For this,
   we use the time-extrapolated solution vector as the convection vector.
   This can be interpreted as a "2nd-order semi-implicit" treatment of the
   convection.

2) Perform 1 discrete-projection-method step as a linear preconditioner
   using a multigrid solver for the linearised Burgers velocity system as
   well as another multigrid solver for the pressure Poisson system.

One can increase the number of iterations for the two methods, however,
there is no "more intelligent" convergence control. The two multigrid
solvers offer more configuration options though -- see the documentation
in the code for details.

Last but not least:
This is a parallel application, of course.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>