added c2d0 and c3d0 mesh files
This commit adds two meshes for the 2D and 3D "flow around a cylinder"
benchmarks, as well as an "auxiliary" 2D mesh.

"c2d0-32-quad.xml" is another mesh defined on the same domain as the
well-known "bench1" mesh. This mesh consists of 32 quadrilateral cells and
has a significantly simpler topology than the "bench1" mesh.

"c3d0-32-quad.xml" is a mesh representing the XY-plane of the 3D flow
around a cylinder benchmark. Note that the (XY-plane of the) domain is
slightly different from the 2D case:
* the channel length is 2.5 instead of 2.2
* the circle midpoint is (0.5, 0.2) instead of (0.2, 0.2)

Due to these differences, also the topology of this mesh is slightly
different: one "vertical slice" of elements has been moved to the left of
the circle.

One purpose of this mesh is to act as input for the mesh-extruder tool,
which can be used to generate the corresponding 3D hexahedral mesh.

Finally, the "c3d0-128-hexa.xml" mesh is a hexahedral mesh for the 3D
benchmark. This file has been generated using the mesh-extruder tool from
the corresponding previous 2D mesh.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>