added check for empty partition in MeshNode::extract_patch
In some rare (and non-deterministic) cases, partitioners as ParMETIS have
the bad habid of creating partitionings with empty partitions.

This is a deadly hazard for any FE code, so a check in the
PatchMeshPartFactory has been added, which is used by the
MeshNode::extract_patch function by the PartitionedDomainControl class.

If the factory detects an empty partition, it will throw an InternalError
to abort program execution. Note that until now, the assembly of the grid
transfer matrices was usually the first place where working with an empty
partition lead to an application crash.

In other words: applications never had a chance to work on empty
partitions, but now the error message should be clearer in the sense that
the root of the evil (empty partition) is mentioned instead of some local
mass matrix being singular, which is merely a symptom of the empty
partition.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>