Fixed clone from ranged vector error
Added a flag to the base container class to signal wether the containers
memory chunks are owned by itself or are only a slice/range from
another containers memory chunk.

The semantics of a ranged vector clone operation are now well defined as
follows: dvr.clone() creates a new dv which is not a clone but owns its
own memory containing a copy of the dvr's memory range.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>