use alternate sanity check in grid transfer assembly
The assembly of grid transfer matrices includes the inversion of local
mass matrices. It turned out that checking the determinant of the mass
matrices as a sanity check is not a good idea when assembling in single
precision, as the determinant may (legally) become subnormal or even
(umerically) zero. The reason is that the mass matrix entries are of
magnitude h^2 (in 2D). However, the matrix inversion still succeeds in
this case, so the determinant-check produces false alerts.

Instead, we now check whether the Frobenius norm of the matrix product of
the inverted fine mesh mass matrix and the inter-level mass matrix is
normal. The entries of this product matrix should usually be in a moderate
range around zero. However, if the mass matrix inversion failed, the
product matrix will usually contain NaNs, infinities or some other bogus,
so that this sanity check should serve its purpose well...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>