FIX: Global::Matrix now sums up columns(), bytes(), used_elements()
Before, just the values of the underlying container were returned which
is not very useful for distributed containers.

Please note the following:

1) The computation of the global number of rows or colums requires the
creation of a Global::Vector and computing its norm2sqr, so it is
expensive and should be used sparingly. This always computes the sizes
wrt. to the Plain Old Data perspective because everything else is
ambiguous (think the number of columns of a global block matrix).

2) used_elements() really reports the sum of all nonzeros over all
ranks, not the number of nonzeros you would get if you were to merge the
distributed type 0 matrices into one big global matrix. So this contains
the "domain decomposition overhead", but it is also the quantity that
matter.

3) The same is true for bytes(), but it is unambiguous in this case.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>