added distributed file I/O helper class
The new DistFileIO class contains a set of functions for reading and
writing text-mode and binary files in a parallel run without accidentally
lauching a DDoS attack on a cluster's file system.

This class currently offers three sets of functions:

* read_common(...)
This function lets the rank 0 process read a file into a std::stringstream
or BinaryStream object and then broadcasts the stream's content to all
other processes. This function can be used if all processes in an MPI-run
need to read in the same file.

* read_sequence(...)
This function lets each process read in a separate file, whose filenames
are generated from a pattern and the process rank. Each process reads in
its file in a Round-Robin fashion, which may not be the most efficient,
but the simplest variant.

* write_sequence(...)
This function is the counterpart of read_sequence, which lets each process
write to a separate file, whose filenames are generated from a pattern.

For the sequence functions, the filename pattern has to contain a
continuous block of asterisks (*), which is replaced by the zero-padded
rank index of the corresponding process.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>