new partitioned domain control and manual partitioning support
This is a major commit which introduces support for manual partitionings
which are stored in the mesh files. These partitionings can be read in
using the MeshFileReader and selected in an application by using the new
PartiDomainControl.

The class for storing the manual partitions is the Geometry::PartitionSet
class, which is now supported by the MeshFileReader, MeshFileWriter and
MeshExtruder class (templates). Moreover, the mesh-extruder and
mesh-to-vtk tools have been adapted to extrude/output the manual
partitions that have been read in from the input mesh file.

The two mesh files "unit-square-quad.xml" and "c2d0-32-quad.xml" have been
enhanced by sets of manual partitionings. More partitionings for the other
mesh files will be added later.

The new PartiDomainControl superseeds the previous
PartitionedDomainControl. In contrast to its predecessor, this new class
is a "wrapper" around several supported partitioning algorithms, which can
be even selected at runtime by specifying the corresponding command line
parameters for a given application.

Finally, the NavierStokesCP2D, the Poisson-Dirichlet-2D and the
Stokes-Poiseuille-2D apps have been modified to use the new
PartiDomainControl class. Note that this also introduced some changes to
the command line parameters which configure the behaviour of the
partitioner.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>