ENH: r-adaptivity in parallel
The evaluation of the nonlinear functional's gradient requires two
vectors to be type-1, meaning they have to be synchronised. As all
MeshQualityFunctionals are patch local only, this means the prepare()
routine was split into two parts: prepare_pre_sync() and
prepare_post_sync(). prepare() now just calls those two in sequence.

This might not be very pretty, but this way one does not have to worry
about using the objects from kernel/meshopt directly or through the
control layer in control/meshopt.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>