ENH: Added mpi_cout_ordered
This can be used to get all MPI ranks to print one array to std::cout,
ordered by ranks. The implementation is the most inefficient and most
portable, as it gathers all information in one rank which then prints to
std::cout. A construction with multiple MPI barriers only works for
certain MPI implementations.

As this implementation creates a communication bottleneck, it is
intended only for debugging purposes.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>