ENH: Major RumpfFunctional overhaul
This include Maple worksheets and some modifications to nonlinear solvers.

The main difference is that the nonlinear forms and the functional
derivatives are now computed in a variational manner and integrated
numerically over the corresponding reference cell. Previously, the
integrals were evaluated "almost analytically" in Maple. The old
functional are still available, but they are no longer used in the
MeshoptControlFactory.

The transformation to the Rumpf reference cell is now expressed through
a material tensor, which is computed by the corresponding RumpfTrafo and
incorporated into the functional and its derivatives via chain rule.

The cubature rule used for the Q1 functionals is hard coded to a closed
Newton-Cotes formula because those use the cell's vertices as evaluation
points. This is important because the determinant det( nabla R_K )
becomes zero or negative if the cell becomes nonconvex, so the local
functional value gets blown up.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>