implemented new MPI wrapper classes in FEAT::Dist namespace
This commit introduces a new set of MPI wrapper classes encapsulated in
the FEAT::Dist namespace ("Dist" for "distributed"), which superseed the
previous implementations from Util::Comm. Most parts of the kernel as
well as all applications have been adapted to use the new implementation,
with the mesh-optimisation stuff being the one notable exception.

This new implementation prepares FEAT applications for the management of
multiple communicators, which is required for the upcoming implementation
of hierarchical partitioning.

As one of the main consequences, the applications now explicitly create a
Dist::Comm object at the beginning of their main function, which is then
passed on to the PartiDomainControl, which takes care of the rest from the
application point of view.

Furthermore, the synchronisation classes in for scalars, vectors and
matrices in FEAT::Global have been reimplemented from scratch.

One notable change is the re-implementation of the matrix synchronisation
(aka type-0-to-type-1 matrix conversion), which can now handle
incompatible buffers between neighboured patches, which was one serious
"brain bug" in the previous implementation (mea culpa).

Note that the old Util::Comm classes still exist after this commit, as a
few kernel classes are still using them.

All of the new Dist classes provide corresponding (dummy) serial
implementations, which are used if FEAT is configured to build without MPI
support.

For details about the usage of the new Dist classes, see the corresponding
doxygen documentation.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>