ENH: Added set_cofactor routine to TinyAlgebra
This sets one matrix to the cofactor (adjunct) matrix of another matrix
and is not to be confused with the adjoint (or adjugate) of a matrix.

There are specialised implementations for up to 6x6 matrices, while
all other dimensions are handled by a generic implementation that only
works for nonsingular matrices. If someone ever needs the cofactor
matrix of a singular matrix of dimension 7x7 and up, this needs to be
revised.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>