added Tutorial 06: global linear algebra
This new tutorial demonstrates how to set up a global linear algebra
system for the parallel solution of a Poisson PDE on the unit-square
domain by setting up the required objects from the Global namespace
along with the assembly of halo mirrors and gates. This tutorial does
not use any Control classes, but performs all required steps in a
"low-level kernel-only" fashion.

The linear solver used in this tutorial is a global PCG preconditioned
by an additive overlapping Schwarz with a local type-1 ILU
solver/preconditioner.

Note that this tutorial uses the UnitCubePatchGenerator to obtain a
decomposition of the domain and does NOT use any sort of more
sophisticated partitioning.

This tutorial is obviously designed to be compiled and run with MPI,
however, the code can also be compiled without MPI. In the latter case,
the tutorial boils down to a simple "serial" solver, as all classes used
by this tutorial offer corresponding non-MPI counterparts where
necessary.

This commit also includes a few new helper functions for some kernel
classes, most notably the 'local()' function for the Global containers,
which should be more intuitive than the (abused) dereferentiation
operator...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>