split Comm:print() into print() and allprint()
The previous stunt with "root = -1" seemed like a nice idea, but turned
out to be somewhat wicked, therefore the functionality of a "collective
print" has been outsourced to a new function "allprint".

Now we have:

* Dist::Comm::print(std::ostream& os, const String& msg, int root = 0)
Prints 'msg' to 'os' if comm.rank() == root and does nothing otherwise.

* Dist::Comm::allprint(std::ostream& os, const String& msg, int root = 0)
Each process sends 'msg' to the root process, which then prints the
messages of all processes in rank-ascending order. Each output line is
prefixed by the rank of the process that the corresponding output line
came from. This function can be very useful for debugging purposes, but
should not be used (excessively) in "production code" due to its high
communication cost and quasi-sequential execution - unless you want to be
struck down by the mighty fist of Amdahl...

Also fixed a missing include.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>