Removed ProductMatVec and Defect classes
These classes were redundant wrappers for the Axpy class, which is now
directly used by the Matrix containers for their apply methods.
For this purpose, the Axpy methods gained the well known beta parameter
y = alpha*Ax + beta*y
which is nevertheless not yet exported to the user by the containers
apply methods.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>