Removed split local functionals
These were for splitting a hypercube into simplices for evaluationg mesh
quality. However, this was only implemented in 2d and somewhat slow. As
the transformation to a different reference cell is now expressed as
using a different material tensor, the old functionality would have been
difficult to port to the new interface. As it offers no real advantages,
it has been removed.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>