ENH: Reordered everything in Meshopt to get rid of Util::Comm
Some quantities (like the number of cells or the total volume) are
needed for mesh optimisation and were synchronised in the Meshopt
classes (using Util::Comm::allreduce), which are supposed to have no
knowledge of the global world.

Now these quantities get added to an std::set that is used my the
Control::Meshopt classes to synchronise them at certain points. That
means that routines like init() and prepare() were split into
**_pre_sync() and **_post_sync(). Also, the Global::Nonlinearfunctional
gained an init() routine to accomplish this.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>