minor tutorial code tweaks
This commit mostly improves the documentation of the tutorial codes.
Besides that, there are also a few minor changes to the codes:

Tutorial 01:
Removed the exploitation of the fact that the solution is an
eigenfunction of the Laplacian. Now the pre-defined LaplaceFunctional
wrapper is used for the assembly of the right-hand-side vector, which is
a "cleaner" way.

Tutorial 05:
Split the main assembly loop into two distinct loops:
First assemble the matrices and filters, then assemble the grid transfer
operators in a separate loop.

Tutorial 06:
Use the format_string() function of the ScalarErrorInfo instead of
stringifying the object.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>