Added new matrix stock, rewrote solver factory
The matrix stock class provides system/transfer matrices, filters and
multigrid hierarchies with different memory, data and index -types in
local and global scope for the solver factory on demand.
The *_solver_factory applications demonstrate the usage of both in
conjunction with the solver configuration ini files in trunk/data/ini.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>