ENH: Mesh optimisation on coarser levels
This means that the mesh is not optimised on the finest level (as
present in the underlying domain control), but on a coarser level. This
required minor adjustments to the solvers and MeshoptSolverFactory,
which now uses Solver::Multigrid instead of Solver::BasicVCycle. The
application configurations and tests have been updated accordingly.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>