ENH: DuDvFunctionalControl now uses MatrixStock
This has some far reaching consequences. As the MatrixStock supports the
synchronisation of matrices, the DuDvFunctional cannot pretend to be a
matrix any more. This led to splitting the MeshoptSystemLevel into
QuadraticSystemLevel (for quadratic functionals) and
NonlinearSystemLevel (for nonlinear functionals) because it was
impossible to keep the unified interface and member access.

With this, MeshoptSolverFactory is obsolote and the corresponding
functionality was merged into SolverFactory. Note that the
MeshoptPrecondFactory still exists because it contains very specialised
code that is of no general use.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>