major domain control overhaul
This commit is a major revisement of the domain control and all of the
other control classes, which was necessary as part of the preparation
for the support of hierarchical partitioning.

The major revisements in this commit (in addition to the assembly/system
control revisment of the previous commits) are:

* The DomainControl class and its derived classes are now templatised in
  the domain level type instead of only the mesh type.

* The classes derived from the DomainLevel class include the trafo and
  FE space objects in addition to the mesh node. Previously, these were
  stored separately in the assembly levels, which do not exist anymore.

* The DomainControl can now manage a hierarchy of partitions along with
  the corresponding communication layers.

* The levels objects in all control objects and applications are now
  ordered "from-fine-to-coarse". The previous "from-coarse-to-fine"
  order did not make sense in the case of hierarchic partitioning.

Note that the PartiDomainControl is currently *not* able to provide a
hierarchical partitioning -- this functionality will be added lateron.

Moreover, the Multigrid solver currently can *not* work on multi-layer
hierarchies -- this functionality will also be added lateron.

A HierarchUnitCubeDomainControl class has been added which can create
hierarchical partitionings for 2D quad meshes. This class is primarily
intended for development and debugging of the hierarchical code and
should not be used for serious application development (yet).
Note that this class is used in the poisson_neumann_2d and
stokes_vortex_2d apps instead of the simple UnitCubeDomainControl for
testing purposes.

Finally, a hierarch_transfer_test-app has been added to test the
correctness of the grid transfer operators on hierarchical partitioned
mesh hierarchies.

To be continued...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>